Problem: #69 App architecture inefficiencies and overhead
Solution: move and destructured some of page files. #69
