+++ imports/ui/components/moderator/approveWalletImage.html
