+++ imports/ui/components/moderator/moderatorPendingCurrency.html
@@ -1,0 +1,8 @@
+<template name="moderatorPendingCurrency">
+  <div class="card border-secondary mb-3 mb-3" style="display:{{display}};min-width:300px;overflow:hidden;">
+  <div class="card-header" style="height:50px"><h4 class="pull-left">{{currencyName}} ({{currencySymbol}})</h4></div>
+  <div class="card-body">
+    <button type="button" class="btn btn-danger" id="reject">Reject</button><button type="button" class="btn btn-success" id="approve">Approve</button>
+  </div>
+  </div>
+</template>
