+++ imports/ui/components/moderator/moderatorPendingCurrency.js
@@ -1,4 +1,5 @@
 import { Template } from 'meteor/templating'
+import './moderatorPendingCurrency.html'
 
 Template.moderatorPendingCurrency.onCreated(function() {
   // a way to get parent's template instance
@@ -76,20 +77,3 @@ Template.moderatorPendingCurrency.helpers({
 });
 
 
-Template.moderatorPendingAPIBounty.events({
-  'click #approve': function(data) {
-    data.preventDefault();
-    Meteor.call('approveAPIbounty', this._id);
-  },
-  'click #reject': function(data) {
-    data.preventDefault();
-    var reason = $('#reason').val();
-    Meteor.call('rejectBounty', this._id, reason);
-    // Meteor.call('setRejected', this._id, true);
-    // Session.set('currentlyRejecting', this._id);
-    // Session.set('reject', true);
-    // Session.set('submittername', this.username);
-    // Session.set('owner', this.owner);
-    // Session.set('currencyName', this.currencyName);
-  }
-});
