+++ imports/ui/components/ratings/currencyChoice.html
@@ -1,0 +1,4 @@
+<template name="currencyChoice">
+  <button data-toggle="modal" data-target="#upload{{_id}}" id="toggle{{_id}}" type="button" class="btn btn-secondary currencyChoiceBtn" style="width:100%">{{currencyName}} {{{uploadedCount _id}}}
+</button>
+</template>
