+++ imports/ui/components/ratings/currencyChoices.html
@@ -1,0 +1,54 @@
+<template name="currencyChoices">
+  <b>You have already added the following cryptocurrency wallets:</b>
+  <div class="row" style="margin-top: 15px;">
+      <div class="col-sm-4 col-md-4 col-lg-4">
+              <a href="#" class="toggle">{{currencyName}}</a>
+              <div class="row" style="display:none;" id="links_{{_id}}">
+                  <p>Please add a URL to a {{currencyName}} community you are a part of:<br /> 
+                  <input type='text' id='js-com-url_{{_id}}' placeholder="URL here"><button class="js-save">Save</button></p>
+              </div>
+          </div>
+      {{#if subsCacheReady}}
+          {{#each alreadyAdded}}
+              <div class="col-sm-4 col-md-4 col-lg-4">
+                  <a href="/currency/{{slug}}" class="btn btn-secondary" style="width:100%">{{currencyName}}</a>
+              </div>
+          {{else}}
+              {{> empty}}
+          {{/each}}
+      {{else}}
+          {{> loading}}
+      {{/if}}
+  </div>
+  <br />
+  <b>Please select any currencies that you actually use (you must have a wallet installed) and you haven't added previously.</b>
+  <br /> If you don't see your currency, you might want to <a href="/addcoin">add it</a>.
+  <br />
+  <br />
+  <hr>
+  <form id="js-form" class="jumbotron" style="padding:30px;">
+      <b class="lead">Filter by</b>
+      <div class="input-group">
+          <input type="text" id="js-name" class="form-control" placeholder="Enter the name of the cryptocurrency" />
+          <span class="input-group-addon"> OR </span>
+          <input type="text" id="js-symbol" class="form-control" placeholder="Enter the symbol of the cryptocurrency" />
+      </div>
+  </form>
+  <div class="row" style="margin: 15px 0 20px 0;">
+      {{#if subsCacheReady}}
+          {{#each currencies}}
+              <div class="col-sm-4 col-md-4 col-lg-4">
+                  {{> currencyChoice}}    
+              </div>
+              {{> upload}}
+          {{else}}
+              {{> empty}}
+          {{/each}}
+      {{else}}
+          {{> loading}}
+      {{/if}}
+  </div>
+  <div class="row">
+      <button id="populateRatings" class="btn btn-primary">Next</button>
+  </div>
+</template>
