+++ imports/ui/components/ratings/displayRatings.html
@@ -1,0 +1,3 @@
+<template name="displayRatings">
+  {{#with questions}} {{> question}} {{/with}}
+</template>
