+++ imports/ui/pages/ratings.html
@@ -1,0 +1,25 @@
+<template name="ratings">
+    <div class="page-header" id="banner">
+        <div class="row">
+            <div class="col-lg-12">
+                <h1>Ratings</h1>
+            </div>
+        </div>
+    </div>
+    {{#if activeBounty}}
+    <div class="alert alert-danger" role="alert">
+    <h5 class="text-center">{{timeRemaining}}</h5><button class="btn btn-danger" id="js-cancel">Cancel</button></div>
+    {{/if}}
+    {{populateUI}} {{> displayRatings}}
+    <div id="currencyChoices" style="display:none;">
+        {{> currencyChoices}}
+    </div>
+    <div id="displayRatings" style="display:none;" class="text-center">
+        {{> displayRatings}}
+    </div>
+    <!--   <hr />
+  <h5>ADMIN SHORTCUTS (will later be automatic cron jobs)</h5>
+  <button id="elo" class="btn btn-danger">ELO</button>
+  <button id="wallets" class="btn btn-danger">WALLETS</button> -->
+</template>
+
