--- imports/ui/pages/transactions/transactions.template.html
@@ -1,35 +1,0 @@
-<template name="transactions">
-    <h3>Transactions</h3>
-    <table class="table">
-        <thead>
-            <tr>
-                <th scope="col">From</th>
-                <th scope="col">To</th>
-                <th scope="col">Amount</th>
-            </tr>
-        </thead>
-        <tbody>
-            {{#if subsCacheReady}}
-                {{#each transactions}}
-                    <tr>
-                        <td>{{from}}</td>
-                        <td>{{to}}</td>
-                        <td><span style="color: {{color}}">{{amount}}</span> KZR</td>
-                    </tr>
-                {{else}}
-                    {{> empty}}
-                {{/each}}
-            {{else}}
-                {{> loading}}
-            {{/if}}
-        </tbody>
-    </table>
-    <span style="display: flex; align-items: center;justify-content: center;"><a class="btn" href="/transactions/1">|&lt;</a>
-    <a class="btn" href="/transactions/{{prevPage}}">&lt;&lt;</a>
-    {{#each pages}}
-        <a href="/transactions/{{this}}" class="btn {{pageClass}}">{{this}}</a>
-    {{/each}}
-    <a class="btn" href="/transactions/{{nextPage}}">&gt;&gt;</a>
-    <a class="btn" href="/transactions/{{lastPage}}">&gt;|</a></span><br />
-    <h4>Total coins moved from the main wallet: {{total}}</h4>
-</template>
