+++ imports/ui/pages/userProfile.html
@@ -1,0 +1,58 @@
+<template name="userProfile">
+    <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
+        <div class="card-header" style="height:50px">
+            <h4 class="pull-left">
+                {{user.username}}
+            </h4>
+        </div>
+        <div class="card-body">
+            {{#if user.profilePicture}}<img src="{{user.profilePicture}}" style="width: 400px;"/><br /><br />{{/if}}
+            Full name: {{val userData.fullname}}<br />
+            Email address: {{val user.email}}<br />
+            Biography: <pre>{{val user.bio}}</pre><br />
+            Roles: {{val roles}}<br />
+        </div>
+    </div>
+    <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
+        <div class="card-header" style="height:50px">
+            <h4 class="pull-left">
+                Addedd currencies
+            </h4>
+        </div>
+        <div class="card-body">
+            {{#if subsCacheReady}}
+                {{#each currencies}}
+                    <a style="padding: 10px" href="/currency/{{slug}}">{{currencyName}}</a> 
+                {{else}}
+                    {{> empty}}
+                {{/each}}
+            {{else}}
+                {{> loading}}
+            {{/if}}
+        </div>
+    </div>
+    <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
+        <div class="card-header" style="height:50px">
+            <h4 class="pull-left">
+                Comment history
+            </h4>
+        </div>
+        <div class="card-body">
+            {{#if subsCacheReady}}
+                {{#each comments}}
+                    <div class="bigCommentBox">
+                        <div class="commentBody">
+                          <div class="commentText">{{comment}} <small class="text-muted" style="white-space:nowrap;">&#x2015;{{author}}</small>
+                        </div>
+                        </div>
+                          <div class="commentFooter"><div class="pull-left">Posted in <a href="/currency/{{commentMeta.currencySlug}}">{{commentMeta.currency}}</a></div></div>
+                    </div>
+                {{else}}
+                    {{> empty}}
+                {{/each}}
+            {{else}}
+                {{> loading}}
+            {{/if}}
+        </div>
+    </div>
+</template>
