--- imports/ui/pages/userProfile/userHover.template.html
@@ -1,6 +1,0 @@
-<template name="userHover">
-	<!--we can add any data here that should be displayed on hover-->
-	<a href="/profile/{{user.slug}}">{{user.username}}<br />
-	<img src="{{user.profilePicture}}" style="width: 200px;"/>
-	</a>
-</template>
