--- views/templates/normal/currency.html
@@ -1,61 +1,0 @@
-<template name="currency2">
-  <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
-  <div class="card-header" style="height:50px"><h4 class="pull-left"><img style="width:50px; height:50px padding-right:10px" src="{{doesCoinImageExist currencyLogoFilename}}"> {{currencyName}} ({{currencySymbol}})  <i class="fa fa-tachometer" id="dash" aria-hidden="true"></i>
-</h4><h4 class="pull-right">RazorRank: {{razorRank}}</h4></div>
-  <div class="card-body" style="display:{{bodydisplay}};">
-
-
-
-    <div class="row">
-      <div class="col-md-12 col-lg-12">
-        <div class="card">
-          <div class="card-block" style="padding: 10px;">
-            <h5 class="card-title"><i class="fa fa-thumbs-up" style="color:#205482;" aria-hidden="true"></i> The Good</h5>
-            <p class="card-text text-justify">The quick, brown fox jumps over a lazy dog. DJs flock by when MTV ax quiz prog. Junk MTV quiz graced by fox whelps. Bawds jog, flick quartz <i class="fa fa-comments" style="color: #5bc0de;" aria-hidden="true"></i></p>
-          </div>
-        </div>
-      </div>
-      <div class="col-md-12 col-lg-12">
-        <div class="card">
-          <div class="card-block" style="padding: 10px;">
-            <h5 class="card-title"><i class="fa fa-exclamation-triangle" style="color:#FFBC42;" aria-hidden="true"></i> The Bad</h5>
-            <p class="card-text text-justify">Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et ma <i class="fa fa-comments" style="color: #5bc0de;" aria-hidden="true"></i></p>
-          </div>
-        </div>
-      </div>
-    </div>
-<br />
-  <div class="float-left">
-  <ul class="list-group" style="max-width: 300px;">
-  <li class="list-group-item border-secondary"><b>Secured by:</b><br /><i class="fa fa-gavel"></i> {{consensusSecurity}} <a href="#" class="text-info">consensus</a><br /><i class="fa fa-bolt"></i><i class="fa fa-bolt"></i> {{hashpower}} W/s of <a href="#" class="text-info">hashpower</a></li>
-  <li class="list-group-item border-secondary"><b>{{circulating}} {{currencySymbol}}</b> are currently priced at <b>${{price}}</b> each for a total of <b>${{marketCap}}</b> <br />
-    If all future {{currencySymbol}} were on the market today, the price would be <b>${{finalValue}}</b>.
-  </li>
-  <li class="list-group-item border-secondary">
-    <i class="fa fa-clock-o" aria-hidden="true"></i> {{launchDate}} <a href={{link}} class="text-info">{{linktext}}</a></li>
-    <li class="list-group-item border-secondary">Reliability: <p class="text-danger"><i class="fa fa-thumbs-o-down" aria-hidden="true"></i> more data needed</p></li>
-  </ul>
-  </div>
-
-    <div class="float-left"><canvas id="{{_id}}distribution" style="max-width:300px;"></canvas>
-    </div>
-    <div><canvas id="{{_id}}-radar"></canvas></div>
-    <div class="float-left" style="width: 300px; height: 100px;"><a href={{gitRepo}}>github</a></div>
-  </div></div>
-</template>
-
-<template name="currency">
-  <div class="currency-card">
-    <div class="header">
-      <div>{{currencyName}}</div>
-      <div>{{currencySymbol}}</div>
-    </div>
-    <div class="graph">
-      {{> radarGraph _id=_id width=400 height=300 options=graphOptions}}
-    </div>
-    <div class="footer">
-      <div>$180/10m</div>
-      <div>RazorRank: {{razorRank}}</div>
-    </div>
-  </div>
-</template>
